Fixed issue with vertx threads
