+++ src/main/kotlin/fi/metatavu/vp/deliveryinfo/rest/AbstractApi.kt
@@ -1,10 +1,17 @@
 package fi.metatavu.vp.deliveryinfo.rest
 
+import io.smallrye.mutiny.Uni
+import io.smallrye.mutiny.coroutines.asUni
+import io.vertx.core.Context
+import io.vertx.core.Vertx
 import jakarta.inject.Inject
 import jakarta.ws.rs.core.Response
+import kotlinx.coroutines.*
 import org.eclipse.microprofile.config.inject.ConfigProperty
 import org.eclipse.microprofile.jwt.JsonWebToken
+import java.lang.Runnable
 import java.util.*
+import kotlin.coroutines.CoroutineContext
 
 /**
  * Abstract base class for all API services
@@ -206,6 +213,37 @@ abstract class AbstractApi {
         return "$entity with id $id not found"
     }
 
+    /**
+     * Executes a block with coroutine scope
+     *
+     * @param requestTimeOut request timeout in milliseconds. Default is 10000
+     * @param block block to execute
+     * @return Uni
+     */
+    @OptIn(ExperimentalCoroutinesApi::class)
+    fun <T> withCoroutineScope(requestTimeOut: Long = 10000L, block: suspend () -> T): Uni<T> {
+        val context = Vertx.currentContext()
+        val dispatcher = VertxCoroutineDispatcher(context)
+
+        return CoroutineScope(context = dispatcher)
+            .async {
+                withTimeout(requestTimeOut) {
+                    block()
+                }
+            }
+            .asUni()
+    }
+
+    /**
+     * Custom vertx coroutine dispatcher that keeps the context stable during the execution
+     */
+    private class VertxCoroutineDispatcher(private val vertxContext: Context): CoroutineDispatcher() {
+        override fun dispatch(context: CoroutineContext, block: Runnable) {
+            vertxContext.runOnContext {
+                block.run()
+            }
+        }
+    }
 
     companion object {
         const val NOT_FOUND_MESSAGE = "Not found"
